pyproject.toml
[tool.poetry]
name = "conductor-server"
version = "0.1.4"
description = "Server for Conductor"
authors = ["csmizzle <chrissmith700@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
boto3 = "^1.34.98"
conductor = {git = "https://github.com/csmizzle/conductor.git"}
celery = "^5.4.0"
django-redis = "^5.4.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
whitenoise = "^6.6.0"
logfire = {extras = ["celery", "django"], version = "^0.33.0"}
drf-nested-routers = "^0.94.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-django = "^4.8.0"
pre-commit = "^3.7.0"
conductor-discord = {git = "https://github.com/csmizzle/conductor-discord.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
